+++ src/views/CupView/index.js
@@ -1,16 +1,17 @@
 import React from 'react'
-import products from '../../../db.json'
-import ProductCard from '../../../Components/UI/ProductCard'
+import products from '../../db.json'
+import ProductCard from '../../Components/ProductCard'
 import { CardContent } from './styles'
 import { Layout } from 'antd'
 
-const TshirtView = () => {
+const CupView = () => {
   return (
+
     <Layout>
       <CardContent>
         {
           products
-            .filter(prod => prod.type === 'tshirt')
+            .filter(prod => prod.type === 'cup')
             .map(product => {
               return (
                 <ProductCard product={product} key={product.id} />
@@ -22,4 +23,4 @@ const TshirtView = () => {
   )
 }
 
-export default TshirtView
+export default CupView
