+++ src/views/CupView/styles.js
