+++ src/views/TshirtView/index.js
@@ -1,17 +1,16 @@
 import React from 'react'
-import products from '../../../db.json'
-import ProductCard from '../../../Components/UI/ProductCard'
+import products from '../../db.json'
+import ProductCard from '../../Components/ProductCard'
 import { CardContent } from './styles'
 import { Layout } from 'antd'
 
-const CupView = () => {
+const TshirtView = () => {
   return (
-
     <Layout>
       <CardContent>
         {
           products
-            .filter(prod => prod.type === 'cup')
+            .filter(prod => prod.type === 'tshirt')
             .map(product => {
               return (
                 <ProductCard product={product} key={product.id} />
@@ -23,4 +22,4 @@ const CupView = () => {
   )
 }
 
-export default CupView
+export default TshirtView
